# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Momo Wallet API
  description: MTN MoMo-style digital wallet backend. User registration, cash-in via agents, P2P and merchant transfers, pockets (savings), and transaction history.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /v1/accounts:
    get:
      tags: [Accounts]
      summary: List my accounts (pockets) with balances
      parameters:
        - name: X-User-Id
          in: header
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: List of accounts with balance
          content:
            application/json:
              schema:
                type: object
                properties:
                  accounts:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        userId: { type: string }
                        type: { type: string, enum: [MAIN, SAVINGS, SCHOOL_FEES, MERCHANT] }
                        currency: { type: string }
                        balance: { type: number }
        '401':
          description: Missing X-User-Id
  /v1/accounts/pockets:
    post:
      tags: [Accounts]
      summary: Create a pocket (savings or school_fees)
      operationId: createPocket
      parameters:
        - name: X-User-Id
          in: header
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type: { type: string, enum: [SAVINGS, SCHOOL_FEES] }
      responses:
        '201':
          description: Pocket created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  userId: { type: string }
                  type: { type: string }
                  currency: { type: string }
        '401':
          description: Missing X-User-Id
        '400':
          description: Validation failed or pocket type already exists
  /v1/accounts/{accountId}/balance:
    get:
      tags: [Accounts]
      summary: Get balance for an account
      parameters:
        - name: X-User-Id
          in: header
          required: true
          schema: { type: string, format: uuid }
        - name: accountId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Balance
          content:
            application/json:
              schema:
                type: object
                properties:
                  accountId: { type: string }
                  balance: { type: number }
                  currency: { type: string }
        '401':
          description: Missing X-User-Id
        '404':
          description: Account not found
...